# docs/swagger.yml
openapi: 3.0.0
info:
  title: 'Frontastic Catwalk'
  version: 1.0.0
paths:
...
  /api/account/address/new:
    post:
      summary: 'Add a new address to account'
      description: ''
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                _type:
                  type: string
                  pattern: ^\Frontastic\Common\AccountApiBundle\Domain\Address$
                addressId:
                  type: string
                salutation:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                streetName:
                  type: string
                streetNumber:
                  type: string
                additionalStreetInfo:
                  type: string
                additionalAddressInfo:
                  type: string
                postalCode:
                  type: string
                city:
                  type: string
                country:
                  type: string
                state:
                  type: string
                phone:
                  type: string
                isDefaultBillingAddress:
                  type: boolean
                isDefaultShippingAddress:
                  type: boolean
                dangerousInnerAddress: {  }
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: \Frontastic\Catwalk\FrontendBundle\Controller\Account
  /api/account/address/update:
    post:
      summary: 'Update address information'
      description: 'Requires the addressId to be specified in the address to update'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                _type:
                  type: string
                  pattern: ^\Frontastic\Common\AccountApiBundle\Domain\Address$
                addressId:
                  type: string
                salutation:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                streetName:
                  type: string
                streetNumber:
                  type: string
                additionalStreetInfo:
                  type: string
                additionalAddressInfo:
                  type: string
                postalCode:
                  type: string
                city:
                  type: string
                country:
                  type: string
                state:
                  type: string
                phone:
                  type: string
                isDefaultBillingAddress:
                  type: boolean
                isDefaultShippingAddress:
                  type: boolean
                dangerousInnerAddress: {  }
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: \Frontastic\Catwalk\FrontendBundle\Controller\Account
  /api/account/address/remove:
    post:
      summary: 'Remove address information'
      description: 'Requires (only) the addressId to be specified in the address to remove'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                _type:
                  type: string
                  pattern: ^\Frontastic\Common\AccountApiBundle\Domain\Address$
                addressId:
                  type: string
                salutation:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                streetName:
                  type: string
                streetNumber:
                  type: string
                additionalStreetInfo:
                  type: string
                additionalAddressInfo:
                  type: string
                postalCode:
                  type: string
                city:
                  type: string
                country:
                  type: string
                state:
                  type: string
                phone:
                  type: string
                isDefaultBillingAddress:
                  type: boolean
                isDefaultShippingAddress:
                  type: boolean
                dangerousInnerAddress: {  }
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: \Frontastic\Catwalk\FrontendBundle\Controller\Account
  /api/account/address/setDefaultBilling:
    post:
      summary: 'Set an address as default billing address'
      description: 'Requires (only) the addressId to be specified in the address'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                _type:
                  type: string
                  pattern: ^\Frontastic\Common\AccountApiBundle\Domain\Address$
                addressId:
                  type: string
                salutation:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                streetName:
                  type: string
                streetNumber:
                  type: string
                additionalStreetInfo:
                  type: string
                additionalAddressInfo:
                  type: string
                postalCode:
                  type: string
                city:
                  type: string
                country:
                  type: string
                state:
                  type: string
                phone:
                  type: string
                isDefaultBillingAddress:
                  type: boolean
                isDefaultShippingAddress:
                  type: boolean
                dangerousInnerAddress: {  }
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: \Frontastic\Catwalk\FrontendBundle\Controller\Account
...